# users/src/main/resources/users-openapi.yaml
openapi: 3.1.0
info:
  title: ParanoiaX Users API
  description: User profile management, device handling and authentication.
  version: 1.0.0

servers:
  - url: /api
    description: Main API Gateway

tags:
  - name: Authentication & Onboarding
    description: User registration and access token issuance
  - name: Device Management
    description: Device session management and state migration (Transfer Blob)
  - name: Profile Management & Discovery
    description: User search and public ECC key exchange
  - name: Friendships
    description: User contacts, friend requests, and E2EE key synchronization

paths:
  /v1/users/invite:
    post:
      tags: [Authentication & Onboarding]
      summary: Generate an invite token
      description: Generates a temporary token used for creating invite links or QR codes for new users.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200':
          description: Token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'

  /v1/users:
    put:
      tags: [Authentication & Onboarding]
      summary: Register a new user
      description: Primary user onboarding. Exchanges the registration token for a device session and uploads the public ECC key.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '201':
          description: User successfully registered (No Content)
    patch:
      tags: [Profile Management & Discovery]
      summary: Partial profile update
      description: Updates specific user parameters (username, profile).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '204':
          description: Profile updated successfully (No Content)

  /v1/users/auth/challenge:
    post:
      tags: [Authentication & Onboarding]
      summary: Request an authentication challenge
      description: Receives a challenge to verify the user's identity.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - $ref: '#/components/parameters/DeviceIdRequired'
      responses:
        '200':
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'

  /v1/users/auth:
    post:
      tags: [Authentication & Onboarding]
      summary: Authenticate via challenge signature
      description: Verifies client identity via cryptographic signature to issue a JWT session token.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - $ref: '#/components/parameters/DeviceIdRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'

  /v1/users/auth/refresh:
    post:
      tags: [Authentication & Onboarding]
      summary: Refresh JWT token pair
      description: Exchanges a valid refresh token for a new pair of JWT access and refresh tokens.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'

  /v1/users/devices/migrations/{migration_id}:
    put:
      tags: [Device Management]
      summary: Create migration session & get upload URL
      description: Initiated by the primary device after scanning the out-of-band payload. Creates a migration session for the target device and returns an S3 presigned URL for blob upload.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: migration_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMigrationSessionRequest'
      responses:
        '200':
          description: Session created, upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'

  /v1/users/devices/migrations/{migration_id}/status:
    get:
      tags: [Device Management]
      summary: Poll migration session status
      description: Used by the new device to poll the state machine of the migration session.
      parameters:
        - name: migration_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationStatusResponse'
        '404':
          description: Migration session not found (Pending creation by primary device)

  /v1/users/devices/migrations/{migration_id}/complete-upload:
    post:
      tags: [Device Management]
      summary: Confirm blob upload completion
      description: Called by the primary device after successfully uploading the blob to S3. Transitions the session state to READY_FOR_AUTH.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: migration_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Upload confirmed successfully (No Content)

  /v1/users/devices/migrations/{migration_id}/download-url:
    post:
      tags: [Device Management]
      summary: Request S3 presigned URL for blob download
      description: Called by the new device. Verifies the challenge signature and returns a short-lived presigned URL to download the state blob, along with the device signature provided by the primary device.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: migration_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadBlobRequest'
      responses:
        '200':
          description: Presigned URL and device signature retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'

  /v1/users/devices/{device_id}:
    put:
      tags: [Device Management]
      summary: Register a secondary device
      description: Finalizes the migration process by registering the new device.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '204':
          description: Device registered successfully (No Content)
    delete:
      tags: [Device Management]
      summary: Revoke device session
      description: Invalidates a specific device session, preventing it from obtaining new JWT tokens.
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session revoked successfully (No Content)

  /v1/users/{username}:
    get:
      tags: [ Profile Management & Discovery ]
      summary: Search for a user
      description: Searches for users to initiate a chat and retrieve their UUID.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchItem'

  /v1/users/{user_id}/keys:
    get:
      tags: [Profile Management & Discovery]
      summary: Retrieve public ECC keys
      description: Retrieval of public ECC keys for chat initialization
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Array of user devices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserKeysResponse'

  /v1/users/keys:
    post:
      tags: [Profile Management & Discovery]
      summary: Retrieve public ECC keys
      description: Batch retrieval of public ECC keys for chat initialization (ECIES digital envelope) or key rotation procedures.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              name: userIds
              type: object
              required:
                - user_ids
              properties:
                user_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Array of user keys and their devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserKeysResponse'

  /v1/friendships:
    get:
      tags: [Friendships]
      summary: Sync and retrieve friendships
      description: Fetch a paginated list of friendships, including state updates.
      security:
        - BearerAuth: []
      parameters:
        - name: updated_after
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 500
            minimum: 250
            maximum: 1000
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of friendships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableFriendshipResponse'
    post:
      tags: [Friendships]
      summary: Add a new friend
      description: Initiates a friendship request with another user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFriendshipRequest'
      responses:
        '204':
          description: Friendship request created (No Content)

  /v1/friendships/{id}:
    patch:
      tags: [Friendships]
      summary: Update friendship attributes
      description: Modify the encrypted attributes blob for a specific friendship.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFriendshipRequest'
      responses:
        '200':
          description: Friendship updated successfully
    delete:
      tags: [Friendships]
      summary: Delete a friendship
      description: Removes a connection / converts it to a deleted state.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Friendship deleted (No Content)

  /v1/friendships/{id}/accept:
    post:
      tags: [Friendships]
      summary: Accept friendship request
      description: Accepts an incoming friendship request.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Friendship accepted (No Content)

  /v1/friendships/block:
    post:
      tags: [Friendships]
      summary: Block a user
      description: Blocks a specific friendship connection.
      security:
        - BearerAuth: []
      parameters:
        - name: friendId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User blocked (No Content)

  /v1/friendships/{id}/unblock:
    post:
      tags: [Friendships]
      summary: Unblock a user
      description: Restores a previously blocked user connection.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User unblocked (No Content)

  /v1/friendships/keys:
    get:
      tags: [Friendships]
      summary: Retrieve pending keys for friends
      description: Returns pending keys for friends with ACCEPTED status for a specific device.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 250
            maximum: 1000
        - name: offset
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of pending friendship keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendshipKeyResponse'

  /v1/friendships/keys/delete:
    post:
      tags: [Friendships]
      summary: Delete friendship keys
      description: Removes friendship keys by a provided list of IDs.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteFriendshipKeysRequest'
      responses:
        '204':
          description: Keys deleted successfully (No Content)

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    DeviceIdOptional:
      name: Device-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    DeviceIdRequired:
      name: Device-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    InviteResponse:
      type: object
      properties:
        host:
          type: string
        invite_token:
          type: string
        spki_pin:
          type: string

    RegisterUserRequest:
      type: object
      properties:
        username:
          type: string
        invite_token:
          type: string
        identity_key:
          type: string
        device:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            type:
              type: string
              enum: [MOBILE, DESKTOP]
            identity_key:
              type: string
            encryption_key:
              type: string
            device_signature:
              type: string

    ChallengeResponse:
      type: object
      properties:
        challenge:
          type: string
        expires_in:
          type: integer

    AuthRequest:
      type: object
      properties:
        signature:
          type: string
        challenge:
          type: string

    TokenPairResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string

    MigrationStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [WAITING_FOR_UPLOAD, READY_FOR_AUTH, DOWNLOADING]
          description: Current state of the migration session
        challenge:
          type: string
          description: Provided only when status is READY_FOR_AUTH

    CreateMigrationSessionRequest:
      type: object
      properties:
        device_id:
          type: string
          description: UUID of the new device
        identity_key:
          type: string
          description: Public identity signing key (e.g., Ed25519) of the new device
        encryption_key:
          type: string
          description: Public ephemeral encryption key (e.g., X25519) of the new device
        device_signature:
          type: string
          description: Master key signature generated by the primary device authorizing the new device

    PresignedUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
          format: uri

    DownloadBlobRequest:
      type: object
      properties:
        signature:
          type: string
          description: Signature of the challenge string provided in the polling endpoint

    DownloadUrlResponse:
      type: object
      properties:
        download_url:
          type: string
          format: uri
        device_signature:
          type: string
          description: The Master key signature previously uploaded by the primary device

    RegisterDeviceRequest:
      type: object
      properties:
        migration_id:
          type: string
          format: uuid
        signature:
          type: string
          description: Signature of the challenge proving possession of the identity private key
        identity_key:
          type: string
        encryption_key:
          type: string
        device_name:
          type: string
        device_type:
          type: string
        device_signature:
          type: string
          description: Master key signature authorizing this device

    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
        profile:
          type: string

    UserSearchItem:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        profile:
          type: object
          properties:
            data:
              type: string
            version:
              type: integer

    UserKeysResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        identity_key:
          type: string
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceKeys'

    DeviceKeys:
      type: object
      properties:
        device_id:
          type: string
          format: uuid
        identity_key:
          type: string
        encryption_key:
          type: string
        device_signature:
          type: string
        revoked_at:
          type: string
          format: date-time

    AddFriendshipRequest:
      type: object
      required:
        - friend_id
      properties:
        friend_id:
          type: string
          format: uuid
        attributes:
          type: string

    UpdateFriendshipRequest:
      type: object
      properties:
        attributes:
          type: string

    FriendshipResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        friend_id:
          type: string
          format: uuid
        status:
          type: string
        attributes:
          type: string
        created_at:
          type: string
          format: date-time

    PageableFriendshipResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FriendshipResponse'
        meta:
          type: object
          properties:
            has_more:
              type: boolean
            server_time:
              type: integer
              format: int64

    FriendshipKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        friendship_id:
          type: string
          format: uuid
        shared_key:
          type: string

    DeleteFriendshipKeysRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          maxItems: 1000
          items:
            type: string
            format: uuid